requirement.txt
numpy
matplotlib
opencv-python
pillow
tensorflow
pandas
mediapipe
